Fix weapon not being updated when crouching
- caused by me in the first place, but i forgot this small detail :x
- Updated lara structure a bit for enum variables.
- Fixed some variable not using these enum (or using the wrong one !)

+++ TR5Main/Game/misc.cpp
@@ -2,15 +2,17 @@
 #include "misc.h"
 #include "setup.h"
 #include "level.h"
+#include "lara.h"
+#include "draw.h"
 
 short GF(short animIndex, short frameToStart)
 {
-	return short(Anims[animIndex].frameBase + frameToStart);
+	return Anims[animIndex].frameBase + frameToStart;
 }
 
 short GF2(short objectID, short animIndex, short frameToStart)
 {
-	return short(Anims[Objects[objectID].animIndex + animIndex].frameBase + frameToStart);
+	return Anims[Objects[objectID].animIndex + animIndex].frameBase + frameToStart;
 }
 
 CREATURE_INFO* GetCreatureInfo(ITEM_INFO* item)
